ignore more style things in rubocop
I disagree with all but the last two, the latter is something I'd like as a separate
task (it's noisy and I don't always get to documenting stuff right away, but it should
probably show up in my CI and hollar at me for not doing it). The former I disagree with
in tests -- specifically in after blocks, when I'm removing constants unconditionally from
the namespace. Sometimes it's even appropriate in main library code, but then it should
be justified in a comment and have a rubocop:disable on it.